hw5-v0.5

+++ src/controllers/contacts.js
@@ -5,14 +5,13 @@ import {
     postContact,
     patchContact,
     deleteContact,
+    findContactByEmail,
 } from '../services/contacts.js';
 import parsePaginationParams from '../utils/parsePaginationParams.js';
 import parseSortParams from '../utils/parseSortParams.js';
 import parseFilterParams from '../utils/parseFilterParams.js';
 
 export const getAllContactsController = async (req, res, next) => {
-    console.log('CoNTACTS', req.user);
-
     const { page, perPage } = parsePaginationParams(req.query);
     const { sortOrder, sortBy } = parseSortParams(req.query);
     const filter = parseFilterParams(req.query);
@@ -36,12 +35,17 @@ export const getContactByIdController = async (req, res, next) => {
     const { id } = req.params;
 
     const contact = await getContactById(id);
+    console.log('contact', contact);
     if (!contact) {
         next(createHttpError(404, 'Contact not found.'));
         return;
     }
 
-    if (contact.userId.toString() !== req.user._id.toString()) {
+    if (
+        contact.userId !== undefined
+            ? contact.userId.toString()
+            : null !== req.user._id.toString()
+    ) {
         next(createHttpError(403, 'Contact not allowed!'));
         return;
     }
@@ -54,6 +58,13 @@ export const getContactByIdController = async (req, res, next) => {
 };
 
 export const createContactController = async (req, res, next) => {
+    const { email } = req.body;
+    const usedEmail = await findContactByEmail(email);
+    if (usedEmail) {
+        next(createHttpError(409, 'Email in use!'));
+        return;
+    }
+
     const contact = await postContact({ ...req.body, userId: req.user._id });
 
     res.status(201).json({
@@ -67,6 +78,12 @@ export const updateContactController = async (req, res, next) => {
     const { id } = req.params;
 
     const contact = await patchContact(id, req.body);
+
+    if (contact.userId.toString() !== req.user._id.toString()) {
+        next(createHttpError(403, 'Contact not allowed!'));
+        return;
+    }
+
     if (!contact) {
         next(createHttpError(404, 'Contact not found'));
         res.json({ data: { message: 'Contact not found' } });
@@ -81,6 +98,16 @@ export const updateContactController = async (req, res, next) => {
 
 export const deleteContactController = async (req, res, next) => {
     const { id } = req.params;
+    const data = await getContactById(id);
+    if (
+        data !== null
+            ? data.userId.toString()
+            : null !== req.user._id.toString()
+    ) {
+        next(createHttpError(403, 'Contact not allowed!!!'));
+        return;
+    }
+
     const contact = await deleteContact(id);
 
     if (!contact) {
@@ -89,4 +116,3 @@ export const deleteContactController = async (req, res, next) => {
     }
     res.status(204).send();
 };
-``;
